Refactoring services to prepare for use in api
