+++ src/apps/services/constants.py
