+++ src/apps/services/klima_retirements.py
@@ -1,5 +1,4 @@
-from . import S3
-from . import KeyCacheable, final_cached_command, chained_cached_command, single_cached_command
+from . import S3, KeyCacheable, final_cached_command, chained_cached_command, single_cached_command
 
 
 class KlimaRetirements(KeyCacheable):
