+++ src/apps/services/offsets.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 import pandas as pd
 import numpy as np
-from . import S3, Countries, Tokens, constants
-from . import KeyCacheable, chained_cached_command, final_cached_command
+from . import S3, Countries, Tokens, constants, KeyCacheable, chained_cached_command, final_cached_command
 
 
 class Offsets(KeyCacheable):
