+++ src/apps/tco2_dashboard/c3t.py
@@ -2,7 +2,7 @@ from dash import html
 from dash import dcc
 import dash_bootstrap_components as dbc
 from .constants import BETWEEN_SECTION_STYLE
-from .services import Offsets
+from src.apps.services import Offsets
 
 
 def create_content_c3t(fig_pool_pie_chart):
