+++ src/apps/tco2_dashboard/retirement_trends/impl/retirement_trends_chain.py
@@ -6,7 +6,7 @@ from src.apps.tco2_dashboard.retirement_trends.retirement_trends_types \
     import ChartContent, ListData, TopContent
 import dash_bootstrap_components as dbc
 from dash import html, dcc
-from ...services import Offsets, KlimaRetirements, Metrics
+from src.apps.services import Offsets, KlimaRetirements, Metrics
 
 
 class RetirementTrendsByChain(RetirementTrendsInterface):
