+++ src/apps/tco2_dashboard/retirement_trends/impl/retirement_trends_pool.py
@@ -12,7 +12,7 @@ from src.apps.tco2_dashboard.retirement_trends.retirement_trends_types \
 import dash_bootstrap_components as dbc
 from dash import html, dcc
 import numpy as np
-from ...services import Metrics, KlimaRetirements
+from src.apps.services import Metrics, KlimaRetirements
 
 
 class RetirementTrendsByPool(RetirementTrendsInterface):
